Initial commit.
The extension integrates an ``after_request`` function for a Flask application to execute a defined function whenever a specific status code from response occurs. The extension is useful in case it is necessary to perform an action whenever a status code occurs (e.g. sending a mail notification for internal server error codes).